requirements.txt
asyncio-mqtt==0.16.1
click==8.1.3
uvloop==0.21.0
sipyco @ git+https://github.com/m-labs/sipyco.git@v1.4
ThorlabsPM100
numpy==1.26.2
pyserial==3.3
configobj==5.0.9
apscheduler==3.10.4
pydantic==2.7.3
pyyaml==6.0.1
networkx==3.4.2
quantnet-mq
